GCP_cloud_functions/lake_to_data_warehouse/requirements.txt
google-cloud-storage
google-cloud-pubsub